評価カラムの追加

+++ db/migrate/20201017140435_create_posts.rb
@@ -2,7 +2,7 @@ class CreatePosts < ActiveRecord::Migration[6.0]
   def change
     create_table :posts do |t|
       t.string :title
-      t.float :rate
+      t.integer :rate
 
       t.timestamps
     end
